Implemented Trajectories on the state machine
Trajectories are constructed as an array for each joint, and position commands are generated as interpolated between array elements. Speed of movement can be modified according to TIME parameters (SITSTANDTIME or STEPTIME).

+++ Sit_stand/canopend/src/CO_master.h
@@ -36,7 +36,7 @@
 
 
 extern pthread_mutex_t CO_CAN_VALID_mtx; /* from main */
-extern volatile uint16_t CO_timer1ms;    /* from main */
+extern volatile uint32_t CO_timer1ms;    /* from main */
 
 
 /**
